test: share Yoast stubs and helpers
Replace the Mockery overload mock with a guarded Schema_Types stub so the filter_graph tests are not order-dependent. Drop the extra JSON round-trip assertion. Move duplicated helpers into the unit TestCase.

Addresses PR feedback. Refs #1360
