+++ tests/Unit/Integrations/YoastPresentationPostTest.php
@@ -11,7 +11,6 @@ namespace Automattic\CoAuthorsPlus\Tests\Unit\Integrations;
 
 use Automattic\CoAuthorsPlus\Tests\Unit\TestCase;
 use Brain\Monkey\Filters;
-use Brain\Monkey\Functions;
 use CoAuthors\Integrations\Yoast;
 
 /*
@@ -45,28 +44,6 @@ final class YoastPresentationPostTest extends TestCase {
 		$presentation->context->post = $post;
 
 		return $presentation;
-	}
-
-	/**
-	 * Let get_coauthors() run to completion without WordPress.
-	 *
-	 * With no co-author terms and guest authors forced, it skips both the term
-	 * lookup results and the post_author fallback (so no $wpdb is needed) and goes
-	 * straight to its `get_coauthors` filter, which the calling test asserts on.
-	 */
-	private function stub_get_coauthors_dependencies(): void {
-		Functions\when( 'cap_get_coauthor_terms_for_post' )->justReturn( array() );
-
-		$GLOBALS['coauthors_plus'] = (object) array( 'force_guest_authors' => true );
-	}
-
-	/**
-	 * Clean up the global seeded by stub_get_coauthors_dependencies().
-	 */
-	protected function tear_down(): void {
-		unset( $GLOBALS['coauthors_plus'] );
-
-		parent::tear_down();
 	}
 
 	/**
